File input can be as strange as possible
I fixed the command line to where I forgot to add quotations around the
input and output files. This is fixed now and anything goes. File output
also works, I accidently hardcoded it for my desktop but now it will be
stored to anyones desktop. Eventually I will add a feature to select
output for now this goes.